services/dasboard_api/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
influxdb-client
